Blender/pyproject.toml
[project]
name = "io_scene_ueformat"
version = "0.0.0"
requires-python = ">=3.10"

[project.optional-dependencies]
dev-b4-1 = [  # Python 3.11
    "ruff==0.4.3",
    "numpy==1.24.3",
    "zstandard==0.16.0",
    "fake-bpy-module",  # currently there's no 4.1 version (https://github.com/nutti/fake-bpy-module/discussions/199)
]
dev-b4-0 = [  # Python 3.10
    "ruff==0.4.3",
    "numpy==1.23.5",
    "zstandard==0.16.0",
    "fake-bpy-module-4.0",
]

[tool.pyright]
reportInvalidTypeForm = "none"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN101",
    "ANN102",
    "D",
    "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["io_scene_ueformat"]
